Separo funciones de groups a módulo aparte
