+++ ckanext/gobar_theme/helpers/__init__.py
@@ -1,3 +1,4 @@
 # pylint: disable=wildcard-import
 from .helpers import *
 from .organizations import *
+from .groups import *
